requirements.txt
openai
xhtml2pdf
python-dotenv
